% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_famic_data.R
\name{create_famic_data}
\alias{create_famic_data}
\title{Create FAMIC database}
\usage{
create_famic_data(path, recursive = TRUE)
}
\arguments{
\item{path}{Path to the folder containing coded .xlsm files}

\item{recursive}{Boolean specifying whether sub-folders should also be searched}
}
\value{
A dataframe object
}
\description{
This function searches for .xlsm files in the specified folder (and
optionally sub-folders), and checks whether they are all named in a
ID_WEEKNUMBER_weeks.xlsm format, and whether they contain the appropriate number
of codes. If these checks are passed, the function then extracts all data
from the coded .xlsm files, and organises them in a tibble object.
}
